Added Start of rendering system. Started Vulkan support. I forgot I was working with a oob language. Added inheritance in some areas and removed unnesessary void calls.

+++ Engine/src/modules/engine/Input/IInputSystem.cpp
@@ -4,7 +4,7 @@
 
 #include "IInputSystem.h"
 
-#include "Logger.h"
+#include "../Library/Logger.h"
 
 bool IInputSystem::isKeyDown(const Keys key) {
     if (!bIsInitialized) return false;
@@ -72,7 +72,7 @@ IInputSystem::~IInputSystem() {
     bIsInitialized = false;
 }
 
-void IInputSystem::update(float deltaTime, FF_Memory& ff_memory) {
+void IInputSystem::update(double deltaTime, FF_Memory& ff_memory) {
     if (!bIsInitialized) {
         return;
     }
